+++ CHANGELOG.md
@@ -6,6 +6,18 @@ _“Yesterday I was clever, so I wanted to change the world. Today I am wise, so
 
 
 
+## [0.1.26] - 2021-08-05
+### Report Update
+- Allow flexible return type for any report.compare result
+- Update tests for report.py
+- Update method for testing if a model predicts probabilities in FairCompare
+- Allow user to skip performance measures when using report.compare
+
+
+## [0.1.25] - 2021-08-05
+### Custom Boundary Bug
+- Fix bug that prevented custom boundaries from updating
+
 ## [0.1.24] - 2021-08-03
 ### Undefined Ratios and Fairness Metric Testing
 - Added code to return NaN values in cases where a fairness ratio has a zero denominator